Реализованы REST JSON API для работы с моделями Employee и Position Для доступа по API установлен модуль Guzzle, реализованы соответствующие классы, создана миграция для генерации токена для API.

+++ www/lofty/console/migrations/m211020_070838_add_tables.php
@@ -27,7 +27,7 @@ class m211020_070838_add_tables extends Migration
         $this->createTable(self::TABLE_NAME_POSITIONS, [
             'id' => $this->primaryKey(),
             'name' => $this->string()->notNull(),
-            'salary' => $this->double()->notNull(),
+            'salary' => $this->float()->notNull(),
         ]);
 
         $this->addForeignKey(
